src/third_party/blink/web_tests/fast/forms/file/file-input-webkitdirectory-key-enter-prevent-keypress.html

<!DOCTYPE html>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/choosers/file_chooser.mojom.js"></script>
<script src="../resources/mock-file-chooser.js"></script>
<title>INPUT TYPE=FILE WEBKITDIRECTORY and key events - Enter keypress prevented</title>
<p>
  Ensure pressing Enter key with focus on an INPUT element with
  TYPE=FILE and WEBKITDIRECTORY still launches a file chooser if
  preventDefault() is called on the 'keypress' event. Automated test
  passes if 'FileChooser: opened' was logged.
</p>
<pre id=console></pre>
<input type="file" webkitdirectory>
<script>
document.querySelector('input').onkeypress = e => {
  document.querySelector('#console').innerText = `${e.type} prevented`;
  e.preventDefault();
};

if (testRunner && eventSender) {
  testRunner.dumpAsText();
  testRunner.waitUntilDone();
  mockFileChooserFactory.addEventListener('open', e => testRunner.notifyDone());

  document.querySelector('input').focus();
  // Despite the name, 'keyDown' simulates a full down/press/up sequence.
  eventSender.keyDown('Enter', []);
}
</script>
