src/third_party/blink/web_tests/fast/forms/file/file-input-webkitdirectory-key-enter-prevent-keypress-expected.txt

CONSOLE MESSAGE: line 5: FileChooser: opened; mode=kUploadFolder
Ensure pressing Enter key with focus on an INPUT element with TYPE=FILE and WEBKITDIRECTORY still launches a file chooser if preventDefault() is called on the 'keypress' event. Automated test passes if 'FileChooser: opened' was logged.

keypress prevented
